// types/stripe/invoice-payment-failed-webhook.ts
export interface InvoicePaymentFailed {
    id: string;
    object: string;
    api_version: string;
    created: number;
    data: Data;
    livemode: boolean;
    pending_webhooks: number;
    request: Request;
    type: 'invoice.payment_failed';
}

export interface Data {
    object: DataObject;
}

export interface DataObject {
    id: string;
    object: string;
    account_country: string;
    account_name: string;
    account_tax_ids: null;
    amount_due: number;
    amount_paid: number;
    amount_remaining: number;
    amount_shipping: number;
    application: null;
    application_fee_amount: null;
    attempt_count: number;
    attempted: boolean;
    auto_advance: boolean;
    automatic_tax: AutomaticTax;
    billing_reason: string;
    charge: string;
    collection_method: string;
    created: number;
    currency: string;
    custom_fields: null;
    customer: string;
    customer_address: null;
    customer_email: string;
    customer_name: string;
    customer_phone: null;
    customer_shipping: null;
    customer_tax_exempt: string;
    customer_tax_ids: any[];
    default_payment_method: null;
    default_source: null;
    default_tax_rates: any[];
    description: null;
    discount: null;
    discounts: any[];
    due_date: null;
    ending_balance: number;
    footer: null;
    from_invoice: null;
    hosted_invoice_url: string;
    invoice_pdf: string;
    last_finalization_error: null;
    latest_revision: null;
    lines: Lines;
    livemode: boolean;
    metadata: Metadata;
    next_payment_attempt: null;
    number: null;
    on_behalf_of: null;
    paid: boolean;
    paid_out_of_band: boolean;
    payment_intent: string;
    payment_settings: PaymentSettings;
    period_end: number;
    period_start: number;
    post_payment_credit_notes_amount: number;
    pre_payment_credit_notes_amount: number;
    quote: null;
    receipt_number: null;
    rendering_options: null;
    shipping_cost: null;
    shipping_details: null;
    starting_balance: number;
    statement_descriptor: null;
    status: string;
    status_transitions: StatusTransitions;
    subscription: string;
    subtotal: number;
    subtotal_excluding_tax: number;
    tax: null;
    test_clock: null;
    total: number;
    total_discount_amounts: any[];
    total_excluding_tax: number;
    total_tax_amounts: any[];
    transfer_data: null;
    webhooks_delivered_at: null;
}
// ...
export interface Request {
    id: string;
    idempotency_key: string;
}